we eventually want to replace the `H*` types with this type (`H256` et
al), but that is quite a hefty refactor so it will be done later. we're
adding this type now to make movement integration easier (they have some
interesting ser/de going on)

also fix some clippy lints that slipped through CI somehow

+++ hubble/src/indexer/dummy/dummy.rs
@@ -155,7 +155,7 @@ impl FetcherClient for DummyFetcherClient {
         // sleep(Duration::from_millis(10)).await;
         let reference = BlockReference::new(
             match selection {
-                BlockSelection::LastFinalized => 43000 as u64,
+                BlockSelection::LastFinalized => 43000,
                 BlockSelection::Height(height) => height,
             },
             match selection {
